feat: getUserTickets 메소드명 수정
