+++ application/ticket-app-api/src/test/java/com/uket/app/ticket/api/service/TicketingServiceTest.java
@@ -99,92 +99,4 @@ class TicketingServiceTest {
 
         assertThat(persistReservation.getReservedCount()).isEqualTo(reservation.getTotalCount());
     }
-
-    @Test
-    @Transactional
-    void 유저가_발급한_티켓이_전부_반환된다() {
-        Users user = userRepository.save(Users.builder()
-            .name("test")
-            .role(UserRole.ROLE_USER)
-            .platform(Platform.KAKAO)
-            .platformId("1234")
-            .email("abc@naver.com")
-            .build()
-        );
-
-        University university = universityRepository.save(University.builder()
-            .currentEvent(1L)
-            .emailPostFix("@konkuk.ac.kr")
-            .logoPath("konkuk.png")
-            .name("건국대")
-            .build());
-
-        Events konkuk = eventRepository.save(Events.builder()
-            .university(university)
-            .startDate(LocalDate.now())
-            .endDate(LocalDate.now())
-            .location("건국대 노천극장")
-            .build()
-        );
-
-        Shows konkukFirstShow = showRepository.save(Shows.builder()
-            .totalTicketCount(0)
-            .startDate(LocalDateTime.now())
-            .endDate(LocalDateTime.now())
-            .location("노천극장 건국대")
-            .event(konkuk)
-            .ticketingDate(LocalDateTime.now())
-            .build());
-
-        Shows konkukSecondShow = showRepository.save(Shows.builder()
-            .totalTicketCount(0)
-            .startDate(LocalDateTime.now().plusDays(1))
-            .endDate(LocalDateTime.now().plusDays(1))
-            .location("노천극장 건국대")
-            .event(konkuk)
-            .ticketingDate(LocalDateTime.now())
-            .build());
-
-        Reservation konkukFirstShowReserve = reservationRepository.save(Reservation.builder()
-            .totalCount(10)
-            .reservedCount(0)
-            .show(konkukFirstShow)
-            .type(ReservationUserType.TICKETING_STUDENT)
-            .startTime(LocalDateTime.now())
-            .endTime(LocalDateTime.now())
-            .build());
-
-        Reservation konkukSecondShowReserve = reservationRepository.save(Reservation.builder()
-            .totalCount(10)
-            .reservedCount(0)
-            .show(konkukFirstShow)
-            .type(ReservationUserType.TICKETING_STUDENT)
-            .startTime(LocalDateTime.now())
-            .endTime(LocalDateTime.now())
-            .build());
-
-        Ticket ticket1 = ticketRepository.save(Ticket.builder()
-            .user(user)
-            .event(konkuk)
-            .show(konkukFirstShow)
-            .reservation(konkukFirstShowReserve)
-            .status(TicketStatus.BEFORE_ENTER)
-            .ticketNo(UUID.randomUUID().toString())
-            .build());
-
-        Ticket ticket2 = ticketRepository.save(Ticket.builder()
-            .user(user)
-            .event(konkuk)
-            .show(konkukSecondShow)
-            .reservation(konkukSecondShowReserve)
-            .status(TicketStatus.BEFORE_ENTER)
-            .ticketNo(UUID.randomUUID().toString())
-            .build());
-
-        List<CheckTicketDto> tickets = ticketingService.checkUserTickets(user.getId());
-
-        assertThat(tickets).hasSize(2);
-        assertThat(tickets.get(0).userName()).isEqualTo(user.getName());
-        assertThat(tickets.get(1).userName()).isEqualTo(user.getName());
-    }
 }
